requirements.txt
# Music Embedding Service - Requirements
# For use with Voyage AI API

voyageai>=0.2.0
numpy>=1.20.0
supabase>=2.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.0

# sbase_embeddings_rerun.py --plot-universe (PNG and/or interactive HTML)
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.18.0
